Add language to code blocks where missing
We used to avoid adding `shell` to code blocks because it added
line numbers and looked strange. We no longer add line numbers to
code blocks so this isn't a problem anymore. For some reason,
kramdown renders code blocks with no language specified as
`<pre><code>code</code></pre>` and the padding around the `<code>`
element makes the first line of the code block look indented. This
change should fix that issue.