Fetch all history in sync job
The problem was that we didn't fetch enough history for git to
know how to relate the two branches that we are trying to sync.

By default, the checkout actino only fetch the latest commit on
the current branch, so as far as Git can tell, the checked out
branch is completely unrelated to the remote mirror branch.